xml/parsing/manakai/notations-2.dat

#data
<!DOCTYPE a[
  <!NOTATION c++ SYSTEM "">
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!NOTATION c++ "" "">
| <{} a>

#data
<!DOCTYPE a[
  <!NOTATION c++ SYSTEM "">
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:not name;;c++;m
#document
| <!DOCTYPE a>
|   <!NOTATION c++ "" "">
| <{} a>

#data
<!DOCTYPE a[
  <!NOTATION hoge PUBLIC "`ab" "">
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;xml:pubid:bad char;;`ab;m
#document
| <!DOCTYPE a>
|   <!NOTATION hoge "`ab" "">
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!NOTATION foo PUBLIC "fo>
]><x/>
#errors
2;28;parser:literal not closed;;;m
#document
| <!DOCTYPE x>
|   <!NOTATION foo "fo" "">
| <{} x>
#entities

#data
<?xml version="1.0"?><!DOCTYPE x[
  <!NOTATION foo PUBLIC 'fo>
]><x/>
#errors
2;28;parser:literal not closed;;;m
#document
| <!DOCTYPE x>
|   <!NOTATION foo "fo" "">
| <{} x>
#entities
